feat : 타이틀과 출간일을 문자열로 변환 this book에 문자열로 반환한것과 other book에 문자열로 반환한것이 같은지 확인 hashcode를 이용하여 같은 코드인지 확인 오래된책들로부터 비교하여 정렬
