+++ lib/240315/book.dart
@@ -20,10 +20,6 @@ class Book implements Comparable<Book> {
       other is Book &&
           toString() == other.toString(); //this book에 문자열로 반환한것과 other book에 문자열로 반환한것이 같은지 확인
 
-  // @override
-  // bool operator ==(Object other) =>
-  //     identical(this, other) ||
-  //     other is Book && publishDate == other.publishDate && title == other.title;
 
   @override
   int get hashCode =>
@@ -46,45 +42,6 @@ class Book implements Comparable<Book> {
       publishDate: publishDate ?? this.publishDate,
     );
   }
-}
-
-void main() {
-  final Book book1 = Book(
-      title: '한정은1', comment: '코멘트', publishDate: DateTime.parse('2024-01-03'));
-  final Book book2 = Book(
-      title: '한정은2', comment: '코멘트', publishDate: DateTime.parse('2024-01-01'));
-  final Book book3 = Book(
-      title: '한정은3', comment: '코멘트', publishDate: DateTime.parse('2024-01-02'));
-  final Book book4 = Book(
-      title: '한정은3', comment: '코멘트', publishDate: DateTime.parse('2024-01-02'));
-  final Book book5 = book4; //book5는 book 객체의 book4 인스턴스의 주소를 가지고 있다(?)
-
-  Book book6;
-
-  book4.title = '한정은5';
-  print(book5.title); //공유하고 있다는 코드
-  print(book4.title);
-
-  print(book1 == book2);
-  print(book2 == book3);
-  print(book4.hashCode);
-  print(book5.hashCode); //책이 같은지 같지 않은지 테스트
-  print(book1.publishDate);
-  print(book3.publishDate);
-
-  final booklist = [book1, book2, book3];
-  booklist.sort();
-  for (var book in booklist) {
-    print(book);
-  }
-  print(booklist);
-
-
-  final Book copiedBook = book3.copyWith();
-
-  print(book3); // 출력: Book { title: Title, publishDate: 2024-01-01 00:00:00.000 }
-  print(copiedBook);   // 출력: Book { title: New Title, publishDate: 2024-01-01 00:00:00.000 }
-
 }
 
 //다음 동작을 할 수 있도록 Book 클래스를 수정하시오.
